add QuickPulseEnvelopeFactory
